+++ CHANGELOG.md
@@ -5,11 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## Unreleased
 
 ### Added
 
+- Added endpoints for performing OAuth Device Authorization Grant with Posit Cloud login. (#2692)
 - Added `product_type` field to the config schema. (#2729)
+
+## [1.18.1]
+
+### Fixed
+
+- Updated software components used by Posit Publisher to address CVE-2025-7783
+  (#2741)
 
 ## [1.18.0]