+++ extensions/vscode/CHANGELOG.md
@@ -12,6 +12,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added `product_type` field to the config schema. (#2729)
 
+## [1.18.1]
+
+### Fixed
+
+- Updated software components used by Posit Publisher to address CVE-2025-7783
+  (#2741)
+
 ## [1.18.0]
 
 ### Fixed